refac: move loggedInUser Component to /user, rename [userID] to userComponents
